Prepare app for data schema migrations
