+++ MyLists/Persistence/SwiftData/Extensions/Blueprint+Utils.swift
@@ -1,0 +1,15 @@
+//
+//  MetaList.swift
+//  ReusableLists
+//
+//  Created by Otávio Zabaleta on 26/02/2024.
+//
+
+import Foundation
+import SwiftData
+
+extension Blueprint {
+    static var placeholderBlueprint: Blueprint {
+        Blueprint("Placeholder")
+    }
+}
