+++ MyLists/Persistence/SwiftData/Extensions/BlueprintItem+Utils.swift
@@ -8,19 +8,6 @@
 import SwiftUI
 import SwiftData
 
-@Model
-final class BlueprintItem: ObservableObject {
-    var name: String
-    
-    init(_ name: String) {
-        self.name = name
-    }
-    
-    func asToDoItem() -> ToDoItem {
-        ToDoItem(name)
-    }
-}
-
 extension Array where Element == BlueprintItem {
     var sortedByName: [BlueprintItem] {
         sorted { $0.name < $1.name }
