+++ MyLists/Persistence/SwiftData/Util/NewEntityItem.swift
