+++ MyLists/View/About/AboutView.swift
@@ -28,7 +28,7 @@ struct AboutView: View {
                     Label("Share", systemImage: "square.and.arrow.up")
                 }
                 
-                LabeledContent("Version", value: "2.4.0")
+                LabeledContent("Version", value: "2.4.1")
                 
                 HTMLView(fileName: "PrivacyPolicy")
                     .frame(height: UIScreen.main.bounds.height * 0.7)
